requirements.txt
absl-py==0.9.0
astor==0.8.1
certifi==2019.11.28
cycler==0.10.0
gast==0.2.2
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
importlib-metadata==1.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
Markdown==3.2.2
matplotlib==3.1.3
numpy==1.18.1
opt-einsum==3.2.1
pandas==1.0.1
protobuf==3.12.1
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
rktools @ file:///home/developer/workspace/rklearn-lib/swap/rktools-1.0.2-py2.py3-none-any.whl
setuptools>=41.0.0
six==1.14.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
tornado==6.0.3
tqdm==4.46.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
